Avoid doing a Butler write in DP02_14
Removed an unnecessary round-trip of the injection catalog through the
Butler from tutorial DP02_14.  The ingestion of this catalog was effectively
writing it and reading it back unchanged, so it isn't necessary.

We are about to transition DP02 to use client/server Butler.  Client/server
Butler does not yet support writes.